fix loading the RTT typekit definition on platforms that have unsigned char (ARM)
This is deeply broken, and the proper fix should be in typelib. For now,
workaround
