--- lib/orogen/ocl.orogen
@@ -1,16 +1,0 @@
-name "OCL"
-
-context = external_task_context "ReportingComponent" do
-    fixed_initial_state
-end
-
-external_task_context 'TCPReporting' do
-    subclasses "ReportingComponent"
-    property 'port', 'int', 3142
-end
-
-external_task_context 'FileReporting' do
-    subclasses "ReportingComponent"
-    property 'ReportFile', 'std/string'
-end
-
