--- lib/orogen/orocos.tlb
@@ -1,13 +1,0 @@
-<?xml version="1.0"?>
-<typelib>
-  <numeric name="/int8_t" category="sint" size="1" />
-  <numeric name="/int64_t" category="sint" size="8" />
-  <numeric name="/uint64_t" category="uint" size="8" />
-  <null name="/nil" />
-  <alias name="/int64" source="/int64_t"/>
-  <alias name="/uint64" source="/uint64_t"/>
-  <container  name="/std/string" of="/int8_t" size="0" kind="/std/string" >
-  <metadata key="orogen_include"><![CDATA[:string]]></metadata>
-  </container>
-  <alias name="/string" source="/std/string"/>
-</typelib>
