--- lib/orogen/orocos.typelist
@@ -1,4 +1,0 @@
-/string
-/uint
-/int64
-/uint64
