--- lib/orogen/rtt.orogen
@@ -1,15 +1,0 @@
-name "RTT"
-
-external_task_context "TaskContext" do
-    abstract
-    root_model
-
-    rtt_states = ['INIT', 'PRE_OPERATIONAL', 'FATAL_ERROR', 'EXCEPTION', 'STOPPED',
-        'RUNNING', 'RUNTIME_ERROR'].
-        map { |v| [v, :toplevel] }
-
-    # Define the standard RTT states, but hack to *not* enable the extended
-    # state support
-    instance_variable_set :@states, rtt_states
-end
-
